# ARROW-17305: [C++] Avoid spending time in popcount in BitmapAnd benchmark