@westonpace 2022-08-04
Yep, if I set `ARROW_SIMD_LEVEL=NONE` then I get slower performance (it's actually pretty similar to what your before numbers were) so there must be some kind of auto-vectorization going on.